修复bug (#15)
* 修改部分地方，使代码更规范

* 添加了同步localstorage的插件，更改了分享页面的代码使其可以编辑资源，由于后台数据问题还未跑通，先提交

* 新增更新资源逻辑

* 页面详情封面添加图片

* 修复搜索分页的问题

* 修复登录相关的bug，增加对资源操作的权限检测

* 增加 检测 编辑资源的权限

+++ src/views/movie/RelatedResources/index.vue
@@ -10,7 +10,7 @@
           <footer>
             <span>分享者：{{ item.sharer.name }}</span>
             <time>{{ item.create_at }}</time>
-            <div>
+            <div v-if="item.sharer.name === userName">
               <span
                 @click="editResources(item.mid, item.id, index)"
               >编辑</span>
@@ -41,7 +41,14 @@ export default {
   },
   data () {
     return {
-      data: []
+      data: [],
+      userName: ''
+    }
+  },
+  mounted () {
+    const {accessToken, idToken, userInfo} = this.$store.getters.user
+    if (accessToken && idToken) {
+      this.userName = userInfo.uname
     }
   },
   watch: {
@@ -51,20 +58,38 @@ export default {
   },
   methods: {
     deleteResources (mid, rid) {
-      deleteMovieResources(mid, rid)
-        .then(response => {
-          this.data = this.data.filter(item => item.id !== rid)
+      const {accessToken, idToken} = this.$store.getters.user
+      if (accessToken && idToken) {
+        deleteMovieResources(mid, rid)
+          .then(response => {
+            this.data = this.data.filter(item => item.id !== rid)
+          })
+          .catch(e => console.log(e))
+      } else {
+        this.$notify({
+          title: '操作失败',
+          message: '您没有权限操作！，请先登录',
+          type: 'warning'
         })
-        .catch(e => console.log(e))
+      }
     },
     editResources (mid, rid, index) {
-      this.$router.push({
-        path: `/share/${mid}`,
-        query: {
-          resourcesData: this.data[index],
-          rid: rid
-        }
-      })
+      const {accessToken, idToken} = this.$store.getters.user
+      if (accessToken && idToken) {
+        this.$router.push({
+          path: `/share/${mid}`,
+          query: {
+            resourcesData: this.data[index],
+            rid: rid
+          }
+        })
+      } else {
+        this.$notify({
+          title: '操作失败',
+          message: '您没有权限操作！，请先登录',
+          type: 'warning'
+        })
+      }
     }
   }
 }
